add benchmark code (#74)
* add benchmark code
* add image size
