+++ benchmark/.gitignore
@@ -1,0 +1,3 @@
+latest
+older
+.local
