+++ benchmark/requirements.txt
@@ -1,0 +1,3 @@
+timeout_decorator
+memory_profiler
+six
